requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
Babel==2.12.1
billiard==4.1.0
celery==5.3.4
celery-flower==1.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docker==6.1.3
drf-spectacular==0.26.5
flower==2.0.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanize==4.8.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.5
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==6.5.0
tornado==6.3.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.3
Werkzeug==2.3.7
